fix: Android JSON crash (aimPoint float→Double) and factory reset 403
- Fixture.aimPoint changed from List<Int>? to List<Double>? — server
  returns floats (1239.0) which Kotlin serializer can't parse as Int
- SurfaceTransform pos/rot/scale changed from List<Int> to List<Double>
  for same reason
- Factory reset @Headers("X-SlyLED-Confirm: true") added to Retrofit
  interface — was returning 403 without it
- New JSON model compatibility test (test_json_model_compat) validates
  field types match between server and Android models
- 212 total test assertions, all passing

+++ android/app/src/main/java/com/slywombat/slyled/data/model/Models.kt
@@ -262,7 +262,7 @@ data class Fixture(
     val dmxStartAddr: Int? = null,
     val dmxChannelCount: Int? = null,
     val dmxProfileId: String? = null,
-    val aimPoint: List<Int>? = null,
+    val aimPoint: List<Double>? = null,
     val childIds: List<Int> = emptyList(),
     val strings: List<FixtureString> = emptyList(),
     val rotation: List<Double> = listOf(0.0, 0.0, 0.0),
@@ -309,9 +309,9 @@ data class Surface(
 
 @Serializable
 data class SurfaceTransform(
-    val pos: List<Int> = listOf(0, 0, 0),
-    val rot: List<Int> = listOf(0, 0, 0),
-    val scale: List<Int> = listOf(2000, 1500, 100),
+    val pos: List<Double> = listOf(0.0, 0.0, 0.0),
+    val rot: List<Double> = listOf(0.0, 0.0, 0.0),
+    val scale: List<Double> = listOf(2000.0, 1500.0, 100.0),
 )
 
 // ── Spatial Effects (Phase 3) ──────────────────────────────────────────
